Password encryption working - Breaking whole friends flow as a result
Picking this back up after some time not working on it, I seem to have added
password encryption to the model file and then not tested it. I know that I
was loosely following the code from Miguel Grinberg's Flasky github. It is not
currently working. Now switching to somewhat follow the tutorial from
http://exploreflask.com/en/latest/users.html