Reformatted User class, progress on reformatted entire model.py to change how friends and email validation worked
Initially I was changing the model to allow for email validation. The more I
thought about it the more I wanted to allow for two types of profiles, information
a contact created and information the user stored. The changes are so radical
that I am not saving and testing each change to the Model. As I am going, I am
also changing other things along the way that would be more codeful, accurate.
If the model migrates fine and everything tests fine in postgres, it means rewriting
all the methods and routes that I had already written before progressing. Man, it
pays to thing your sprints through, I have rewritten everything so many times and
will do again before I'm done. That is why this is a learning project and also
a portfolio piece because I want it to really reflect me.