Cleaned up Git and Githhub
Tried to push changes to github and realized that absolutely everything had
changed about this project. Created two branches on github to freeze the
project as it was at graduation and as it was before converting it into a
large app and overhauling the model. Development will now be done in V2 until
deployment on Heroku and then will be merged into master.