See extended commit message
Fixed bug in diet model where the repr was calling something that was no longer
defined in the model.
A bigger problem, I created a new bug when I added a new blueprint that was not in the Miguel Grinberg Flask book: adding a database query in the userdata/forms.py
area of the userdata blueprint caused this error: "RuntimeError: application not registered on db instance and no application bound to current context".
Fixed it with the solution from the comments here, to add db.app = app.
The main advice to use "with" did not help me.
http://stackoverflow.com/questions/19437883/when-scattering-flask-models-runtimeerror-application-not-registered-on-db-w
Also added a Flask_Config environment variable to secrets.sh