Many changes to login and model
I had the wrong bcrypt encrypt and validate on the model. I think I
must have put in the one the Flask book was using.
Added some test people, friends, party to the seed file
Working on the party datetime model.
Have not tested the work I have done today. For Monday or sooner: update
seed file to the new party class methods and test. Then work on getting
the right information in the session and making that clear to Jinja.