Fixed login, logout
When remember_me was selected at Login and you closed the browser, the session
information for user, party and friend was not saved. I added an alternative
session cookie following this information:
http://thecircuitnerd.com/flask-login-tokens/
I was looking to fix a smaller problem, that of how to see what was in the
remember_me session. I found that information but I also found the advice for
alternative tokens to be great.